smaller screen testijng
adjusted the nav on every page when a smaller device is being used so it is not as large as it was before and is more portiontal to the page. tested on chrome and safari with a smaller screen. they all work.

adjusted the header to suit the smaller screen page by having a smaller text size which was tested and successfully works on a samller screen size and chrome and safari.