h2 and nav
changed the h2 to suit the smaller screen by making the font size smaller and repositions the accomdation one as it was not postiotion in the middle.

also adjusted the nav buttons to a smaller size with suficiant space around them so the the user can cleary pick one to tap on.