text size
change text size on small screen to something slighlty lager improve the ease which the users will read it. Also decreased the widths of the area fop text as it looked bad going all the way to the edges and now there is white space for teh user to rest their eyes. tested on chrome on smaller screen. looks good and works